+++ src/main/java/HCMUT/TutorSytem/dto/MajorDTO.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class MajorDTO {
-    private Long id;
+    private Integer id;
     private String name;
     private String majorCode;
     private String programCode;
     private String note;
-    private Long departmentId;
+    private Integer departmentId;
     private String departmentName;
 }
