+++ src/main/java/HCMUT/TutorSytem/payload/request/TutorRequest.java
@@ -7,9 +7,9 @@ import java.util.List;
 @Data
 public class TutorRequest {
     private String title;
-    private Long majorId; // Major determines department
+    private Integer majorId; // Major determines department
     private String description;
-    private List<Long> subjects; // List of subject IDs from frontend
+    private List<Integer> subjects; // List of subject IDs from frontend
     private Integer experienceYears;
     // rating: auto calculated from reviews
     // isAvailable: auto set to true
