+++ src/main/java/HCMUT/TutorSytem/repo/TutorScheduleRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TutorScheduleRepository extends JpaRepository<TutorSchedule, Long> {
-    List<TutorSchedule> findByTutorId(Long tutorId);
+public interface TutorScheduleRepository extends JpaRepository<TutorSchedule, Integer> {
+    List<TutorSchedule> findByTutorId(Integer tutorId);
 }
