+++ src/main/java/HCMUT/TutorSytem/service/UserService.java
@@ -3,5 +3,5 @@ package HCMUT.TutorSytem.service;
 import HCMUT.TutorSytem.model.Datacore;
 
 public interface UserService {
-    Long getInfoFromHcmutSystem(Datacore data);
+    Integer getInfoFromHcmutSystem(Datacore data);
 }
